Move modelEvaluate function to its own file
