+++ src/Interpreters/CatBoostModel.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <Interpreters/IExternalLoadable.h>
 #include <Columns/IColumn.h>
 #include <Columns/ColumnsNumber.h>
